AABB-Particle Collisions
When an object is created you can call initAABB which will generate a
bounding box using the positions and scaling of the objects instances.
To make this bounding box collidable with the particle affect you call
particleEffect->addCollidable(theObject). Once the object has been given
to the particle effect it will check, at every update, to see if any of
the particles are colliding with the objects. When they are colliding
the particle will bounce and its lifespan set to a smaller 'bounceValue'
which times it so it dies shortly after it bounces.